+++ examples/multiFileProject/multiFileProject.h
@@ -12,5 +12,11 @@
 
 #pragma once
 
+// Use larger queue size if necessary for large data transfer. Default is 512 bytes if not defined here
+#define ASYNC_QUEUE_LENGTH     512
+
+// Use larger priority if necessary. Default is 10 if not defined here. Must be > 4 or adjusted to 4
+#define CONFIG_ASYNC_TCP_PRIORITY   (12)
+
 // Can be included as many times as necessary, without `Multiple Definitions` Linker Error
 #include "AsyncTCP_SSL.hpp"